feat: Add basic Go code examples
This commit adds several Go code examples to the repository. The examples cover topics such as constants, for loops, if-else statements, switch statements, values, and variables. Each code file contains a simple implementation of the respective concept.

The added code files are:
- constants.go
- for.go
- go.mod
- helloworld.go
- ifelse.go
- switch.go
- values.go
- variables.go

These examples will serve as a starting point for beginners to learn and understand the basics of Go programming.